[RENAME FILES] refactored script to rename HNS to Markdown files in place
